+++ source/particle.h
@@ -215,7 +215,7 @@ public:
 
 	PYTHON() void printParts(IndexInt start=-1, IndexInt stop=-1, bool printIndex=false);
 	PYTHON() void clearFile(std::string filename = "data.txt");
-	PYTHON() void getCurrentData(std::string filename, std::string title, int cfl, int resolution, FlagGrid& flags, MACGrid &vel, bool lastFrame=false);
+	PYTHON() void getCurrentData(std::string filename, std::string title, float cfl, int resolution, FlagGrid& flags, MACGrid &vel, bool lastFrame=false);
 
 	//! get data pointer of particle data
 	PYTHON() std::string getDataPointer();
